requirements.txt
pyperclip>=1.8.2
requests>=2.28.1
msal>=1.20.0
azure-servicebus>=7.8.1
sense-hat>=2.4.0